Apple Pay (#88)
* Add Apple Pay

* Fix code style

* Print apple pay token data

* Apple Pay: encrypt and send order ID in request application data

* Update code style

* Fix apple pay localization

* Send apple pay token to server

* Update code style

+++ Chisto/Data/Model/PromoCode.swift
@@ -12,7 +12,7 @@ import ObjectMapper
 class PromoCode: Mappable {
   var id: Int = 0
   var code: String = ""
-  var discount: Int = 0
+  var discount: Decimal = 0
 
   func mapping(map: Map) {
     id <- map["id"]
